Remove slight pop & make sound always update
When initializing a song then fading in, a pop may be heard because the sound is always set to max when loading a song.
This is fixed by writing to DevSound's sound volume instead of directly to NR50.
However this means that NR50 isn't updated when just init'ing a song, so NR50 is always updated (not just when fading).
Plus this lets software directly change DevSound's volume control and it being actually applied.

Also slight 1-byte optimization. Wee.